refactor: fold getBoundingBox() into measureElement()
Address review feedback: instead of a separate getBoundingBox() API,
extend measureElement() to return {x, y, width, height}. This keeps
one measurement concept in ink rather than three overlapping APIs.
