+++ src/index.ts
@@ -36,8 +36,7 @@ export type {BoxMetrics, UseBoxMetricsResult} from './hooks/use-box-metrics.js';
 export {default as useBoxMetrics} from './hooks/use-box-metrics.js';
 export type {CursorPosition} from './log-update.js';
 export {default as measureElement} from './measure-element.js';
-export type {BoundingBox} from './get-bounding-box.js';
-export {default as getBoundingBox} from './get-bounding-box.js';
+export type {ElementMetrics} from './measure-element.js';
 export type {DOMElement} from './dom.js';
 export {kittyFlags, kittyModifiers} from './kitty-keyboard.js';
 export type {KittyKeyboardOptions, KittyFlagName} from './kitty-keyboard.js';
